+++ R/power_scABEL1.R
@@ -121,7 +121,7 @@ power.scABEL1 <- function(alpha=0.05, theta1, theta2, theta0, CV, n,
             formatC(ptm["elapsed"], digits=2), "\n")
     #print(ptm)
     # return the vector of all counts
-    names(p) <- c("p(BE)", "p(BE-wABEL)", "p(BE-pe)", "p(BE-ABE)")
+    names(p) <- c("p(BE)", "p(BE-ABEL)", "p(BE-pe)", "p(BE-ABE)")
     if (!pe_constr) p <- p[-3] # without pe constraint
     p
   } else {