+++ NEWS.md
@@ -1,6 +1,6 @@
 # PowerTOST 1.4-7.9000
 
-Published on GitHub 2019-08-16
+Published on GitHub 2019-08-20
 
 ## Bug fixes
 
@@ -21,6 +21,8 @@ Published on GitHub 2019-08-16
 
 ## Minor changes
 
+  * Output of `sampleN.TOST()`: Same name of replicate designs like in the scaled functions.
+  * In `power.scABEL()` name of scaled component `"p(BE-ABEL)"` instead of `"p(BE-wABEL)"`.
   * Updated `test_ABEL.R` in `inst/tests` to support subject simulations.
   * `power.scABEL()`: More informative warning about heteroscedasticity in the partial replicate design (use of `power.scABEL.sdsims()` suggested).
   * Add/subtract `.Machine$double.eps` if `rho` is -1 or +1 given in `power_type1_2TOST.R` (similar to `sampleN_2TOST_sim.R`). Removed warning in the latter function. Less confusing for users and the example in the man-page looks nicer.